requirements.txt
Django==1.6.2
South==0.8.4
beautifulsoup4==4.3.2
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
httplib2==0.8
oauth2==1.5.211
oauthlib==0.6.1
psycopg2==2.5.2
pystache==0.5.3
python-openid==2.2.5
python-social-auth==0.1.21
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
sorl-thumbnail==11.12
static==1.0.2
wsgiref==0.1.2
